+++ lib/jazzy/gem_version.rb
@@ -1,3 +1,3 @@
 module Jazzy
-  VERSION = '0.0.19' unless defined? Jazzy::VERSION
+  VERSION = '0.0.20' unless defined? Jazzy::VERSION
 end